requirements.txt
torch
torchaudio
numpy
python-dotenv
soundfile
matplotlib
tqdm